# docs/09-api/openapi.yaml
openapi: 3.1.0
info:
  title: Lumine API
  description: Institutional AI-native quantitative hedge fund API
  version: 0.1.0
paths:
...
  /api/v1/rpc/run-decision-cycle:
    post:
      tags:
      - rpc
      summary: Run Decision Cycle
      description: Enqueue a full decision-cycle workflow run.
      operationId: run_decision_cycle_api_v1_rpc_run_decision_cycle_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RpcCommandRequest'
              - type: 'null'
              title: Request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rpc/halt-trading:
    post:
      tags:
      - rpc
      summary: Halt Trading
      description: Halt all trading activity at the operational level.
      operationId: halt_trading_api_v1_rpc_halt_trading_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rpc/resume-trading:
    post:
      tags:
      - rpc
      summary: Resume Trading
      description: Resume trading activity after an operational halt.
      operationId: resume_trading_api_v1_rpc_resume_trading_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rpc/cancel-order:
    post:
      tags:
      - rpc
      summary: Cancel Order
      description: Request cancellation of an open order.
      operationId: cancel_order_api_v1_rpc_cancel_order_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RpcCommandRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcCommandResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rpc/commands/{command_id}:
    get:
      tags:
      - rpc
      summary: Get Command Status
      description: Return the execution status/result of a previously enqueued command.
      operationId: get_command_status_api_v1_rpc_commands__command_id__get
      parameters:
      - name: command_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Command Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RpcCommandResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RpcCommandRequest:
      properties:
        command:
          type: string
          title: Command
        parameters:
          additionalProperties: true
          type: object
          title: Parameters
      type: object
      required:
      - command
      title: RpcCommandRequest
      description: Generic RPC command payload.
    RpcCommandResponse:
      properties:
        command_id:
          type: string
          format: uuid
          title: Command Id
        command:
          type: string
          title: Command
        status:
          type: string
          enum:
          - accepted
          - rejected
          title: Status
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        enqueued_at:
          type: string
          format: date-time
          title: Enqueued At
      type: object
      required:
      - command_id
      - command
      - status
      - enqueued_at
      title: RpcCommandResponse
      description: RPC command acceptance receipt.
    RpcCommandResult:
      properties:
        command_id:
          type: string
          format: uuid
          title: Command Id
        command:
          type: string
          title: Command
        status:
          type: string
          enum:
          - queued
          - processing
          - completed
          - failed
          title: Status
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        enqueued_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Enqueued At
        processed_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Processed At
      type: object
      required:
      - command_id
      - command
      - status
      title: RpcCommandResult
      description: RPC command execution status/result (polled by command_id).
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...